% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{bandwith.scaling}
\alias{bandwith.scaling}
\title{Scaling Factor for the Local Bandwidth}
\usage{
bandwith.scaling(tau)
}
\arguments{
\item{tau}{Numeric, representing the extremile level, which is used to tailor the
scaling factor. This should be a value between 0 and 1.}
}
\value{
Numeric, the computed scaling factor for the local bandwidth. This factor
        is used to adjust bandwidth found via cross-validation.
}
\description{
Computes the scaling factor for the local bandwidth based on a given extremile level.
}
\examples{
tau_value <- 0.9
scaling_factor <- bandwith.scaling(tau = tau_value)
print(scaling_factor)

}
\references{
Battagliola, M. L., and Bladt, M. (2024).
Extremile scalar-on-function regression with application to climate scenarios.
\emph{Under review}.

Daouia, A., Gijbels I. and Stupfler, G. (2022).
Extremile regression.
\emph{Journal of the American Statistical Association} 117 1579–1586.
}
\author{
Maria Laura Battagliola
}
