% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bandwith_F.R
\name{optband.F}
\alias{optband.F}
\title{Optimal Bandwidth Calculation for Conditional CDF Estimator}
\usage{
optband.F(X, s, Y, n.h = 20, kernCDFI = 3, kappa = 1)
}
\arguments{
\item{X}{Matrix of functional input data where each row represents an observation
and each column represents the evaluation of the function on a certain point of the grid.}

\item{s}{A vector indicating the common grid on which the functions are evaluated}

\item{Y}{Response variable corresponding to the rows of X.}

\item{n.h}{Number of bandwidths to consider in the optimization process.
Default is 20.}

\item{kernCDFI}{Kernel choice indicated by an integer used in the difference
function calculation. Default is 3, corresponding to Epanechnikov}

\item{kappa}{A scaling factor which controls the trade-off between estimated bias and variance. Default is 1.}
}
\value{
A vector of optimal bandwidths, one for each observation in the dataset.
}
\description{
Optimal Bandwidth Calculation for Conditional CDF Estimator
}
\examples{
# Simulating some data
n <- 200
S <- 100
tau <- 0.5
simData <- simData(n=n, tau=tau, S=S)
X <- simData[[2]]$X
Y <- simData[[2]]$Y

# Calculating optimal bandwidths for the conditional CDF
s <- seq(0,1,length.out=S)
n.h <- 10
optimal_bandwidths <- optband.F(X, s, Y, n.h)
print(optimal_bandwidths)


}
\references{
Battagliola, M. L., and Bladt, M. (2024).
Extremile scalar-on-function regression with application to climate scenarios.
\emph{Under review}.

Chagny, G., and Roche, A. (2014).
Adaptive and minimax estimation of the cumulative distribution function given a functional covariate.
\emph{Electronic Journal of Statistics} 8 2352 – 2404.
}
\author{
Maria Laura Battagliola
}
