% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simData.R
\name{simData}
\alias{simData}
\title{Simulate Functional Data for Extremile Regression Models}
\usage{
simData(n, tau, gamma = 0, S, myseed = 1234, sd.eps = 0.25)
}
\arguments{
\item{n}{Integer, the number of observations to simulate.}

\item{tau}{Numeric, extremile level, used in the calculation of certain model components.}

\item{gamma}{Numeric, the heteroskedasticity parameter, defaults to 0.}

\item{S}{Integer, the number of points in regular grid over [0,1] to evaluate the functions over.}

\item{myseed}{Integer, seed for random number generation to ensure reproducibility, defaults to 1234.}

\item{sd.eps}{Numeric, standard deviation of the error term, defaults to 0.25.}
}
\value{
A list containing:
\itemize{
  \item \code{tau} Extremile level used in the simulation.
  \item \code{data} Data frame containing the response variable \code{Y}, the true extremile \code{xi},
        and the matrix of regressors \code{X}.
  \item \code{grid} The sequence of points in the functional coordinate used to construct \code{X}.
}
}
\description{
Simulate Functional Data for Extremile Regression Models
}
\details{
The function constructs the functional data \code{X} by combining sinusoidal basis functions
with random coefficients. The response \code{Y} is modeled as the integral of the product
of \code{X} and a functional coefficient \code{beta(s)} plus an error term that is scaled
by both a random component and the \code{gamma} parameter.
}
\examples{
simulated_data <- simData(n = 100, tau = 0.9, gamma = 0.5, S = 200)
par(mfrow=c(1,2))
plot(simulated_data$data$Y, simulated_data$data$xi, main = "Response vs. Extremile")
abline(a=0, b=1)
matplot(simulated_data$grid,t(simulated_data$data$X),type="l", main="Functional covariates")

}
\references{
Battagliola, M. L., and Bladt, M. (2024).
Extremile scalar-on-function regression with application to climate scenarios.
\emph{Under review}.
}
\author{
Maria Laura Battagliola
}
