% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimStudy.R
\name{SimStudy}
\alias{SimStudy}
\title{Conduct a Simulation Study for Extremile Regression}
\usage{
SimStudy(
  n,
  S,
  gamma,
  tau.gen,
  sd.eps = 0.25,
  tau.est,
  percent = 0.95,
  B,
  my_seed,
  hFMat = NULL,
  comparison.fllr = FALSE,
  kappa = 1,
  n.hF = 20
)
}
\arguments{
\item{n}{Integer, the number of observations to simulate in each dataset.}

\item{S}{Integer, the number of points in regular grid over [0,1].}

\item{gamma}{Numeric, the heteroskedasticity parameter.}

\item{tau.gen}{Numeric, the extremile level used for generating data.}

\item{sd.eps}{Numeric, standard deviation of the error term, defaults to 0.25.}

\item{tau.est}{Numeric, the extremile level used for estimation in the simulation.}

\item{percent}{Numeric, the percentage of variance explained when computing the eigenfunctions, defaults to 0.95.}

\item{B}{Integer, the number of simulation iterations.}

\item{my_seed}{Integer, seed for random number generation to ensure reproducibility.}

\item{hFMat}{Matrix, optional, pre-specified bandwidths for each iteration. If NULL,
bandwidths are calculated within the function. Defaults to NULL.}

\item{comparison.fllr}{Logical, indicates whether to perform a comparison with
functional linear regression. Only conducted when \code{tau.est} is 0.5. Defaults to FALSE.}

\item{kappa}{Numeric, a scaling factor used in the bandwidth selection, defaults to 1.}

\item{n.hF}{Integer, the number of points for the bandwidth function, defaults to 20.}
}
\value{
A list containing:
\itemize{
  \item \code{True extremiles} Matrix of true extremiles used in data generation.
  \item \code{Estimated extremiles} Matrix of extremiles estimated from the simulated data.
  \item \code{MSE} Vector of mean squared errors calculated for each simulation.
  \item \code{CDF bandwidths} Matrix of calculated bandwidths for each simulation.
  \item \code{Estimated means} (Optional) Matrix of estimated means from functional linear regression,
        included only if \code{comparison.fllr} is TRUE and \code{tau.est} is 0.5.
}
}
\description{
This function performs a simulation study to assess the effectiveness of extremile
functional regression. It generates datasets, computes extremiles, and optionally compares
these to results obtained from functional linear regression when certain conditions are met.
}
\details{
The function iteratively generates data using \code{simData}, then applies \code{ExtrFunReg}
to estimate extremiles. If specified, it also compares these estimates to those obtained via
functional linear regression (\code{fllr}) when \code{tau.est} is 0.5. This function is useful
for evaluating the performance of estimation methods under varying simulation conditions.
}
\examples{
results <- SimStudy(n = 50, S = 100, gamma = 0.5, tau.gen = 0.9, tau.est = 0.9, B = 10, my_seed = 123)
print(results$MSE)

}
\references{
Battagliola, M. L., and Bladt, M. (2024).
Extremile scalar-on-function regression with application to climate scenarios.
\emph{Under review}.
}
\author{
Maria Laura Battagliola
}
